iData: fixed bugs related to the Error propagation for unary operators, and those for which the error formula can not be accurately computed. fixed other bugs revealed from extended tests.

+++ trunk/@iData/isfinite.m
@@ -8,12 +8,12 @@ function a = isfinite(a)
 % output: b: array (int)
 % ex:     b=isfinite(a);
 %
-% Version: $Revision: 1.3 $
+% Version: $Revision: 1.4 $
 % See also iData, iData/sign, iData/isreal, iData/isfinite, iData/isnan,
 %          iData/isinf, iData/isfloat, iData/isinterger,
 %          iData/isnumeric, iData/islogical, iData/isscalar, 
 %          iData/isvector, iData/issparse
 
 a = iData_private_unary(a, 'isfinite');
-a=uint8(a);
+a=uint8(double(a));
